requirements.txt
streamlit
opencv-python
pillow
numpy
scikit-learn
joblib
pandas
